+++ src/helics/core/CoreBroker.hpp
@@ -116,6 +116,9 @@ class CoreBroker: public Broker, public BrokerBase {
     gmlc::concurrency::DelayedObjects<std::string> ActiveQueries; //!< holder for active queries
     JsonMapBuilder fedMap; //!< builder for the federate_map
     std::vector<ActionMessage> fedMapRequestors; //!< list of requesters for the active federate map
+    JsonMapBuilder currentTimeMap; //!< builder for the current time graph
+    std::vector<ActionMessage>
+        ctimeRequestors; //!< list of requesters for the active current time status
     JsonMapBuilder depMap; //!< builder for the dependency graph
     std::vector<ActionMessage> depMapRequestors; //!< list of requesters for the dependency graph
     JsonMapBuilder dataflowMap; //!< builder for the dependency graph
@@ -303,6 +306,8 @@ class CoreBroker: public Broker, public BrokerBase {
     void processError(ActionMessage& command);
     /** disconnect a broker/core*/
     void disconnectBroker(BasicBrokerInfo& brk);
+    /** mark this broker and all other that have this as a parent as disconnected*/
+    void markAsDisconnected(global_broker_id brkid);
     /** run a check for a named interface*/
     void checkForNamedInterface(ActionMessage& command);
     /** remove a named target from an interface*/
@@ -339,6 +344,8 @@ class CoreBroker: public Broker, public BrokerBase {
     void initializeDependencyGraph();
     /** generate a json string containing the data flow information for all federation object*/
     void initializeDataFlowGraph();
+    /** generate a json string containing the time information for the federation */
+    void initializeCurrentTimeMap();
 
     /** send an error code to all direct cores*/
     void sendErrorToImmediateBrokers(int error_code);