+++ src/helics/core/FederateState.cpp
@@ -1657,6 +1657,9 @@ std::string FederateState::processQueryActual(const std::string& query) const
             return std::to_string(dep.baseValue());
         });
     }
+    if (query == "current_time") {
+        return timeCoord->printTimeStatus();
+    }
     if (query == "timeconfig") {
         std::ostringstream s;
         s << "{\n" << timeCoord->generateConfig();
@@ -1691,7 +1694,7 @@ std::string FederateState::processQuery(const std::string& query) const
         qstring = processQueryActual(query);
     } else if ((query == "queries") || (query == "available_queries")) {
         qstring =
-            "publications;inputs;endpoints;interfaces;subscriptions;dependencies;timeconfig;config;dependents";
+            "publications;inputs;endpoints;interfaces;subscriptions;dependencies;timeconfig;config;dependents;current_time";
     } else { // the rest might to prevent a race condition
         if (try_lock()) {
             qstring = processQueryActual(query);