+++ src/helics/core/ForwardingTimeCoordinator.hpp
@@ -119,5 +119,7 @@ class ForwardingTimeCoordinator {
 
     /** check if there are any active Time dependencies*/
     bool hasActiveTimeDependencies() const;
+    /** get the current next time*/
+    Time getNextTime() const { return time_next; };
 };
 } // namespace helics