+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Increased code coverage and additional bug fixes.  The error propagation in HELI
 -  `helics_terminate_on_error` flag to escalate what would be a local error into a global one that will halt the co-simulation.  This flag can be specified through the flag to federates or to brokers and cores through a command line option `--terminate_on_error` 
 -   `addDependency` function was added to the C++ Federate API and shared library API, it can add a direct dependency between federates manually.  
 -   A 32-bit Windows zip install archive for releases
+-   "global_time", "current_time", and "state" queries for brokers and cores, and "current_time" query for federates.
 -   Support for a 'helics-release-build' event trigger to the release build GitHub Actions workflow
 
 ### Deprecated